src/requirements.txt
argon2-cffi~=21.3.0
PyYAML~=6.0
telebot==0.0.5
psycopg2~=2.9.6
